+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-11T05:06:47.477Z",
+    "2026-06-11T09:48:18.486Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {